Add watcher thread to stop worker if is stuck in connecting
In addition to running the existing thread to watch/restart task threads, it will now watch for the ZeebeAdapter connectivity state, and stop the worker if state doesn't change from being in a retrying connection state (after 60 sec by default).

Add prefix to log message, to indicate which watcher thread is logging.